# backend/swagger.yaml
openapi: 3.0.3
info:
  title: API Documentation
  description: API for a service platform with authentication, categories, ads, admin, chat, complaints, job requests, listings, notifications, offers, payments, reviews, settings, users, and verification
  version: 1.0.0
servers:
  - url: /api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter a valid JWT token (access token for most endpoints, or refresh token for `/auth/refresh`) in the format `Bearer <your-token>`.
        For `/auth/refresh`, use the refresh token obtained from `/auth/login` in the `Authorization` header.
        In Swagger UI, click the "Authorize" button, enter the token (without the "Bearer" prefix), and it will be applied to the endpoint.
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
        email:
          type: string
        name:
          type: object
          properties:
            first:
              type: string
            last:
              type: string
        phone:
          type: string
        avatarUrl:
          type: string
        roles:
          type: array
          items:
            type: string
        isPremium:
          type: boolean
        isTopRated:
          type: boolean
        isVerified:
          type: boolean
        totalJobsCompleted:
          type: number
    Category:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        description:
          type: string
        isActive:
          type: boolean
    Ad:
      type: object
      properties:
        _id:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        targetUrl:
          type: string
        duration:
          type: number
        advertiserId:
          $ref: '#/components/schemas/User'
...
    Complaint:
      type: object
      properties:
        _id:
          type: string
        reporterId:
          $ref: '#/components/schemas/User'
        reportedUserId:
          $ref: '#/components/schemas/User'
        jobRequestId:
          type: string
        problemType:
          type: string
          enum: ['late', 'no_show', 'incomplete_work', 'poor_quality', 'rude_behavior', 'price_dispute', 'other']
        description:
          type: string
        status:
          type: string
          enum: ['pending', 'investigating', 'resolved', 'dismissed']
    JobRequest:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        seeker:
          $ref: '#/components/schemas/User'
        assignedTo:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: ['open', 'assigned', 'completed', 'cancelled']
        budget:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
...
    Notification:
      type: object
      properties:
        _id:
          type: string
        userId:
          $ref: '#/components/schemas/User'
        type:
          type: string
        message:
          type: string
        isRead:
          type: boolean
...
    Review:
      type: object
      properties:
        _id:
          type: string
        reviewer:
          $ref: '#/components/schemas/User'
        reviewedUser:
          $ref: '#/components/schemas/User'
        role:
          type: string
          enum: ['seeker', 'provider']
        rating:
          type: number
        comment:
          type: string
        jobRequest:
          type: string
        serviceListing:
          type: string
    UpgradeRequest:
      type: object
      properties:
        _id:
          type: string
        userId:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: ['pending', 'accepted', 'rejected']
        adminExplanation:
          type: string
        rejectionComment:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  value:
                    type: string
        timestamp:
          type: string
          format: date-time
paths:
...
  /categories:
    get:
      summary: Get all categories
      tags: [Categories]
      parameters:
        - in: query
          name: includeInactive
          schema:
            type: boolean
          description: Include inactive categories
      responses:
        '200':
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      categories:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create a new category
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                isActive:
                  type: boolean
      responses:
        '201':
          description: Category created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      category:
                        $ref: '#/components/schemas/Category'
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict (category already exists)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /ads:
    post:
      summary: Create a new advertisement
      tags: [Ads]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                title:
                  type: string
                description:
                  type: string
                imageUrl:
                  type: string
                targetUrl:
                  type: string
                duration:
                  type: number
      responses:
        '201':
          description: Ad created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Ad'
                  message:
                    type: string
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/upgrade-requests:
    get:
      summary: List all upgrade requests
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: ['all', 'pending', 'accepted', 'rejected']
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Upgrade requests retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      requests:
                        type: array
                        items:
                          $ref: '#/components/schemas/UpgradeRequest'
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /complaints:
    post:
      summary: Submit a new complaint
      tags: [Complaints]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reportedUserId:
                  type: string
                jobRequestId:
                  type: string
                problemType:
                  type: string
                  enum: ['late', 'no_show', 'incomplete_work', 'poor_quality', 'rude_behavior', 'price_dispute', 'other']
                description:
                  type: string
      responses:
        '201':
          description: Complaint submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      complaint:
                        $ref: '#/components/schemas/Complaint'
                  message:
                    type: string
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /admin/complaints:
    get:
      summary: Get complaints for admin
      tags: [Complaints]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: ['all', 'pending', 'investigating', 'resolved', 'dismissed']
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Complaints retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      complaints:
                        type: array
                        items:
                          $ref: '#/components/schemas/Complaint'
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /requests:
    post:
      summary: Create a new job request
      tags: [JobRequests]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                category:
                  type: string
                budget:
                  type: object
                  properties:
                    min:
                      type: number
                    max:
                      type: number
                deadline:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Job request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      jobRequest:
                        $ref: '#/components/schemas/JobRequest'
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get all job requests with filtering
      tags: [JobRequests]
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: minBudget
          schema:
            type: number
        - in: query
          name: maxBudget
          schema:
            type: number
        - in: query
          name: lat
          schema:
            type: number
        - in: query
          name: lng
          schema:
            type: number
        - in: query
          name: radius
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Job requests retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      jobRequests:
                        type: array
                        items:
                          $ref: '#/components/schemas/JobRequest'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalCount:
                            type: integer
                          totalPages:
                            type: integer
                          hasNext:
                            type: boolean
                          hasPrev:
                            type: boolean
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /notifications:
    get:
      summary: List notifications for the current user
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Notifications retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          totalPages:
                            type: integer
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /reviews:
    get:
      summary: Get reviews for a specific user or listing
      tags: [Reviews]
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: listingId
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      reviews:
                        type: array
                        items:
                          $ref: '#/components/schemas/Review'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          totalPages:
                            type: integer
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /settings:
    get:
      summary: Get user settings
      tags: [Settings]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      settings:
                        type: object
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update user settings
      tags: [Settings]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notifications:
                  type: object
                  properties:
                    email:
                      type: boolean
                    push:
                      type: boolean
                privacy:
                  type: object
                  properties:
                    showPhone:
                      type: boolean
                    showEmail:
                      type: boolean
      responses:
        '200':
          description: Settings updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      settings:
                        type: object
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users:
    get:
      summary: Search users
      tags: [Users]
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          totalPages:
                            type: integer
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...